services/indicator-service/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pandas==2.1.1
numpy==1.26.0
sqlalchemy==2.0.21
asyncpg==0.28.0
pydantic-settings>=2.2.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
